move AbstractFileClassLoader to compiler
there doesn't seem to be anything repl-specific about it, and it's used in a few places
in the compiler
