+++ compiler/src/dotty/tools/AbstractFileClassLoader.scala
@@ -11,7 +11,6 @@
  */
 
 package dotty.tools
-package repl
 
 import scala.language.unsafeNulls
