+++ compiler/src/dotty/tools/dotc/util/ClasspathFromClassloader.scala
@@ -5,7 +5,7 @@ import scala.language.unsafeNulls
 import java.net.URLClassLoader
 import java.nio.file.Paths
 
-import dotty.tools.repl.AbstractFileClassLoader
+import dotty.tools.AbstractFileClassLoader
 
 object ClasspathFromClassloader {
